Update and test viterbi code
I am going to try switching to logs of probabilities to hopefully gain
precision. When testing the imputation of U6202-080 on chromosome 1B
there were two paths that appeared to be equally likely. This is
something to consider. I don't think that the viterbi algorithm will
be guaranteed to return the most likely path if at a certain stage
there are two partial paths of equal probability. It will still return
a maximal path, but we should consider if that is what we actually
want, or if we would like to have all maximal paths returned. My guess
is that a single maximal path will be sufficient, because even if
there was a maximum path with a second path that was a small epsilon
worse we would not be returning that path even as epsilon approaches 0.