Full vcf imputation working
An entire vcf file can now be imputed. There is still some work to do
in order to get the format back to something that can be saved as an
output vcf. Currently the output is a matrix of 1, 2, 3, NA which
indicates respectively if a site is homozygous parent 1, homozygous
parent 2, heterozygous, or unknown.