Work on progress output
Currently, a table will be printed, but there is no information given
regarding how much progress has actually been made. I will be looking
at using parallel:::mcfork() to start a process to monitor progress by
use of a FIFO buffer/file/pipe. See https://stackoverflow.com/
questions/27726134/
how-to-track-progress-in-mclapply-in-r-in-parallel-package
for an example.