Work on imputation process
I am working on testing if the path returned by the viterbi algorithm
in the forward direction has the same probability as the one in the
reverse direction. In theory they should both be optimal, though they
may not be the same. If this is the case it will be helpful that we
don't have to run the viterbi algorithm multiple times, but we also
need to consider what to do about mutliple paths with equal liklihods