Remove excess functions; add code to clean input
Various function related to running the viterbi algorithm multiple
times have been removed as we now know it is not necessary to run the
viterbi algorithm more than once. The genotype and allic depth
gathering part of the VCF function have been switched to for loops so
that bad sites can be removed.