Update viterbi related function for soft imputation
Soft imputation refers to the fact that the viterbi is now looking for
all optimal paths and reconciling them. For instance, if some optimal
paths suggest that site X should be homozygous parent 1 and other
optimal paths suggest that site X should be heterozygous then the site
is imputed as being a site with one allele from parent 1 and the other
allele is unknown. The code appears to be working well.